+++ parallel_mars.hpp
@@ -8,7 +8,7 @@
 template <typename T>
 auto cuda_mars(SquareMatrix<T> &_J_mat,
                std::vector<T> &_h,
-               size_t _n,
+               int _n,
                int _t_min,
                int _t_max,
                T _c_step,
@@ -27,7 +27,7 @@ auto cuda_mars(SquareMatrix<T> &_J_mat,
 template <typename T>
 auto parallel_mars(SquareMatrix<T> &_J_mat,
                    std::vector<T> &_h,
-                   size_t _n,
+                   int _n,
                    int _t_min,
                    int _t_max,
                    T _c_step,
